requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.6
Flask==2.3.2
flask-restx==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pyserial==3.5
pytz==2023.3
referencing==0.30.2
rpds-py==0.9.2
Werkzeug==2.3.6
